refactor(fe): tidy CLI route layout
- Settings page components → settings/components/ subdir to match the
  convention used elsewhere (e.g. recovery/components/).
- cli/cliAuthorize.ts → cli/utils.ts, matching the existing route util
  naming (access/utils.ts, callback/utils.ts). CLI_GENERIC_DERIVATION_ORIGIN
  moves with it so the constant lives next to its only consumer.
- CLI e2e fixture: narrow server.address() with a guard instead of an
  `as AddressInfo` cast. Note that `listen(0, ...)` already gives each
  fixture instance a fresh OS-picked port.
