+++ src/frontend/src/routes/(new-styling)/manage/(authenticated)/settings/components/CliConfirmDialog.svelte
