+++ src/frontend/src/routes/(new-styling)/manage/(authenticated)/settings/components/VerifyPanel.svelte
